+++ components/nimbus/src/nimbus_client.rs
@@ -2,7 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::enrollment::DisqualifiedReason;
 use crate::{
     behavior::EventStore,
     client::{create_client, SettingsClient},
@@ -326,11 +325,7 @@ impl NimbusClient {
                     is_enrolled_set.insert(ee.slug.clone());
                     all_enrolled_set.insert(ee.slug.clone());
                 }
-                EnrollmentStatus::WasEnrolled { .. }
-                | EnrollmentStatus::Disqualified {
-                    reason: DisqualifiedReason::NotSelected | DisqualifiedReason::NotTargeted,
-                    ..
-                } => {
+                EnrollmentStatus::WasEnrolled { .. } | EnrollmentStatus::Disqualified { .. } => {
                     all_enrolled_set.insert(ee.slug.clone());
                 }
                 _ => {}